commit all

+++ Services/CharacterService.cs
@@ -8,6 +8,8 @@ using donet_rpg.Dtos.Character;
 using donet_rpg.Models;
 using dotnet_rpg.Models;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
+using System.Security.Claims;
 
 namespace donet_rpg.Services
 {
@@ -17,19 +19,21 @@ namespace donet_rpg.Services
 
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAcessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAcessor)
         {
-           
+            _httpContextAcessor = httpContextAcessor;
+
             _mapper = mapper;
-            _context= context;
+            _context = context;
         }
 
-
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllcharacters(int userId)
+        private int GetUserId()=> int.Parse(_httpContextAcessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllcharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.Characters.Where(c=>c.User.Id== userId).ToListAsync();
+            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
@@ -37,7 +41,11 @@ namespace donet_rpg.Services
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c=>c.Id==id);
+            Character dbCharacter = await _context.Characters
+            .Include(c=>c.CharacterSkills).ThenInclude(CookieSecurePolicy=>CookieSecurePolicy.Skill)
+            .Include(c=>c.Weapon)
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id== GetUserId());
+            
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
 
             return serviceResponse;
@@ -47,36 +55,43 @@ namespace donet_rpg.Services
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character ch = _mapper.Map<Character>(newCharacter);
-         await  _context.AddAsync(ch);
-         await _context.SaveChangesAsync();
-           
+            ch.User = await _context.Users.FirstOrDefaultAsync(u => u.Id==GetUserId());
+            await _context.AddAsync(ch);
+            await _context.SaveChangesAsync();
 
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+
+            serviceResponse.Data = (_context.Characters.Where(c=>c.User.Id==GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updatedCharacter)
         {
-           
+
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-               
-                    character.Name = updatedCharacter.Name;
-                    character.Class = updatedCharacter.Class;
-                    character.Defense = updatedCharacter.Defense;
-                    character.HitPoints = updatedCharacter.HitPoints;
-                    character.Intelligence = updatedCharacter.Intelligence;
-                    character.Strength = updatedCharacter.Strength;
+                Character character = await _context.Characters.Include(c=>c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if(character.User.Id == GetUserId())
+                {
+                character.Name = updatedCharacter.Name;
+                character.Class = updatedCharacter.Class;
+                character.Defense = updatedCharacter.Defense;
+                character.HitPoints = updatedCharacter.HitPoints;
+                character.Intelligence = updatedCharacter.Intelligence;
+                character.Strength = updatedCharacter.Strength;
 
-                    _context.Characters.Update(character);
-                    await _context.SaveChangesAsync();
+                _context.Characters.Update(character);
+                await _context.SaveChangesAsync();
 
-                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(_context.Characters.ToListAsync());
-                }
-             
-            
+                serviceResponse.Data = _mapper.Map<GetCharacterDto>(_context.Characters.ToListAsync());
+            }
+            else
+            {
+                serviceResponse.Success=false;
+                serviceResponse.Message="Character Not Found";
+            }
+            }
+
             catch (Exception ex)
             {
                 serviceResponse.Success = false;
@@ -91,11 +106,18 @@ namespace donet_rpg.Services
             try
             {
 
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-               _context.Characters.Remove(character);
-               await _context.SaveChangesAsync();
-                serviceResponse.Data = ( _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                Character character = await _context.Characters.FirstAsync(c => c.Id == id && c.User.Id==GetUserId());
+                if(character!= null)
+                {
+                _context.Characters.Remove(character);
+                await _context.SaveChangesAsync();
+            
+                serviceResponse.Data = (_context.Characters.Where(c=>c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
 
+                }
+                else{
+                    serviceResponse.Message="Character not found";
+                }
             }
             catch (Exception ex)
             {
